PPU: OAM data structure conversion
This commit contains a bunch of commented code but is functional at this
time. The OAM has been converted to a structure of arrays in the hopes
of making things more efficient. There's certainly more optimization to
be had, but at this point I've shaved ~2ms off frame times on my system.